print diagonal slash

+++ 0x04-more_functions_nested_loops/7-print_diagonal.c
@@ -13,7 +13,7 @@
  */
 void print_diagonal(int n)
 {
-	int i, j;
+	signed long int i, j;
 
 	if (n > 0)
 	{
@@ -24,7 +24,6 @@ void print_diagonal(int n)
 				if (i == j)
 				{
 					_putchar('\\');
-					/*_putchar(' ');*/
 				}
 				else if (i > j)
 				{
